+++ test/Spec.hs
@@ -21,3 +21,16 @@ spec = describe "unit tests" $ do
     it "eval changedSince" $ do
       evalSearchExpr (changedSince (read "2021-04-01 00:00:00 UTC"))
         `shouldBe` [("chfieldfrom", Just "2021-04-01T00:00:00Z")]
+
+    it "eval changedRange" $ do
+      evalSearchExpr
+        ( changedRange
+            (read "2021-04-01 00:00:00 UTC")
+            (read "2021-04-13 00:00:00 UTC")
+        )
+        `shouldBe` [ -- TODO: fix unused f1 and f4 query args
+                     ("f1", Just "OP"),
+                     ("f4", Just "CP"),
+                     ("chfieldfrom", Just "2021-04-01T00:00:00Z"),
+                     ("chfieldto", Just "2021-04-13T00:00:00Z")
+                   ]