+++ src/Web/Bugzilla/RedHat.hs
@@ -53,6 +53,7 @@ module Web.Bugzilla.RedHat
 , sendBzRequest
 , intAsText
 
+, Request
 , BugId
 , AttachmentId
 , CommentId